+++ CHANGELOG.md
@@ -5,9 +5,12 @@
 - Fixed a bug where record patterns using the spread operator (`..`) to discard
   unwanted arguments would not type check correctly when the record had no
   labelled fields.
+- Add support for using sized binary segments in pattern matches when targeting
+  JavaScript.
 - A warning is now emitted for double unary negation on ints (`--`) and bools
   (`!!`) as this does nothing but return the original value.
 - The errors emitted when a name is reused in a module have been made clearer.
+- Running a module now uses the dependency's target and runtime in its `gleam.toml`.
 
 ## v0.28.3 - 2023-04-17